+++ milkman-cassandra/src/main/java/milkman/plugin/cassandra/UrlUtil.java
@@ -1,0 +1,31 @@
+package milkman.plugin.cassandra;
+
+import lombok.SneakyThrows;
+import org.apache.commons.lang3.StringUtils;
+
+import java.net.URI;
+import java.net.URLDecoder;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+public class UrlUtil {
+
+
+
+
+    @SneakyThrows
+    public static Map<String, String> splitQuery(URI url)  {
+        Map<String, String> query_pairs = new LinkedHashMap<String, String>();
+        String query = url.getQuery();
+        if (StringUtils.isBlank(query)){
+            return Collections.emptyMap();
+        }
+        String[] pairs = query.split("&");
+        for (String pair : pairs) {
+            int idx = pair.indexOf("=");
+            query_pairs.put(URLDecoder.decode(pair.substring(0, idx), "UTF-8"), URLDecoder.decode(pair.substring(idx + 1), "UTF-8"));
+        }
+        return query_pairs;
+    }
+}