+++ milkman-cassandra/src/main/java/milkman/plugin/cassandra/export/CqlshExporter.java
@@ -1,0 +1,10 @@
+package milkman.plugin.cassandra.export;
+
+import milkman.plugin.cassandra.domain.CassandraRequestContainer;
+import milkman.ui.plugin.AbstractTextExporter;
+
+public class CqlshExporter extends AbstractTextExporter<CassandraRequestContainer> {
+    public CqlshExporter() {
+        super("Cqlsh", new CqlshTextExport());
+    }
+}