+++ milkman-grpc/src/main/java/milkman/plugin/grpc/export/GrpcurlExporter.java
@@ -1,0 +1,10 @@
+package milkman.plugin.grpc.export;
+
+import milkman.plugin.grpc.domain.GrpcRequestContainer;
+import milkman.ui.plugin.AbstractTextExporter;
+
+public class GrpcurlExporter extends AbstractTextExporter<GrpcRequestContainer> {
+    public GrpcurlExporter() {
+        super("gRPCurl", new GrpcurlTextExport());
+    }
+}