+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/curl/CurlExporter.java
@@ -1,6 +1,9 @@
 package milkman.ui.plugin.rest.curl;
 
-public class CurlExporter extends AbstractTextExporter {
+import milkman.ui.plugin.AbstractTextExporter;
+import milkman.ui.plugin.rest.domain.RestRequestContainer;
+
+public class CurlExporter extends AbstractTextExporter<RestRequestContainer> {
 	public CurlExporter() {
 		super("Curl", new CurlTextExport());
 	}