+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/curl/HttpExporter.java
@@ -1,6 +1,9 @@
 package milkman.ui.plugin.rest.curl;
 
-public class HttpExporter extends AbstractTextExporter {
+import milkman.ui.plugin.AbstractTextExporter;
+import milkman.ui.plugin.rest.domain.RestRequestContainer;
+
+public class HttpExporter extends AbstractTextExporter<RestRequestContainer> {
 	public HttpExporter() {
 		super("Http", new HttpTextExport());
 	}