+++ milkman/src/main/java/milkman/ui/plugin/TextExport.java
@@ -1,0 +1,8 @@
+package milkman.ui.plugin;
+
+import milkman.domain.RequestContainer;
+
+public interface TextExport<T extends RequestContainer> {
+
+    public String export(boolean isWindows, T container, Templater templater);
+}